PROCESSING-3530 : Testing phase. Move all "generic" numbers to same type (Double). Fix issues found during test suites. Add logs to debug some values. Scenario 1 passing with this commit: All jobs same resource reservations but different priority values; should execute in order of higher priority first.

+++ src/main/java/org/apache/aurora/scheduler/state/dsr/DominantResource.java
@@ -3,9 +3,9 @@ package org.apache.aurora.scheduler.state.dsr;
 public class DominantResource implements Comparable<DominantResource>{
 
     private final String name;
-    private final Float percentageUsage;
+    private final Double percentageUsage;
 
-    public DominantResource(String name, Float percentageUsage) {
+    public DominantResource(String name, Double percentageUsage) {
         this.name = name;
         this.percentageUsage = percentageUsage;
     }
@@ -14,7 +14,7 @@ public class DominantResource implements Comparable<DominantResource>{
         return name;
     }
 
-    public Float getPercentageUsage() {
+    public Double getPercentageUsage() {
         return percentageUsage;
     }
